+++ sgan/GEST_ACT_NAIS/Workflow/Models/DataClass/Arrondissement.cs
@@ -8,12 +8,12 @@ namespace P7GestAct.Models.DataClass
 {
     public class Arrondissement
     {
-        public Int64 ArrondissementID { get; set; }
+        public int ArrondissementID { get; set; }
 
         [Required]
         public String Arrondissement_Name { get; set; }
 
-        public Int64 DepartementID { get; set; }
+        public int DepartementID { get; set; }
         public virtual Departement Departement { get; set; }
         public virtual ICollection<Commune> Communes { get; set; }
         public virtual ICollection<Personne> Personnes { get; set; }
