+++ sgan/GEST_ACT_NAIS/Workflow/Models/DataClass/Centre.cs
@@ -7,11 +7,11 @@ namespace P7GestAct.Models.DataClass
 {
     public class Centre
     {
-        public Int64 CentreID { get; set; }
+        public int CentreID { get; set; }
         public String C_Name { get; set; }
         public String C_Lieux { get; set; }
 
-        public Int64 CommuneID { get; set; }
+        public int CommuneID { get; set; }
         public virtual Commune Commune { get; set; }
         public virtual ICollection<Declaration> Declarations { get; set; }
     }
