+++ sgan/GEST_ACT_NAIS/Workflow/Models/DataClass/Departement.cs
@@ -8,7 +8,7 @@ namespace P7GestAct.Models.DataClass
 {
     public class Departement
     {
-        public Int64 DepartementID { get; set; }
+        public int DepartementID { get; set; }
 
         [Required]
         public String Dept_Name { get; set; }
@@ -16,7 +16,7 @@ namespace P7GestAct.Models.DataClass
         public String Dept_Chef_Lieu { get; set; }
 
 
-        public Int64 RegionID { get; set; }
+        public int RegionID { get; set; }
         public virtual Region Region { get; set; }
         public virtual ICollection<Arrondissement> Arrondissements { get; set; }
     }
