+++ sgan/GEST_ACT_NAIS/Workflow/Models/DataClass/Domicile.cs
@@ -8,7 +8,7 @@ namespace P7GestAct.Models.DataClass
 {
     public class Domicile
     {
-        public Int64 DomicileID { get; set; }
+        public int DomicileID { get; set; }
 
         [Required]
         public String Pays { get; set; }
@@ -19,7 +19,7 @@ namespace P7GestAct.Models.DataClass
         [Required]
         public String Proprietaire { get; set; }
 
-        public Int64 PersonneID { get; set; }
+        public int PersonneID { get; set; }
         public virtual Personne Personne { get; set; }
 
     }
