+++ sgan/GEST_ACT_NAIS/Workflow/Models/DataClass/Langue.cs
@@ -7,11 +7,11 @@ namespace P7GestAct.Models.DataClass
 {
     public class Langue
     {
-        public Int64 LangueID { get; set; }
+        public int LangueID { get; set; }
         public String Langue_Name { get; set; }
         public String Langue_Small_Name { get; set; }
 
-        public Int64 PaysID { get; set; }
+        public int PaysID { get; set; }
         public virtual Pays Pays { get; set; }
     }
 }
